Se ha definico completamente la enum TimeStampFormat para manejar los distintos formatos de TimeStamp y simplificar su creación. También se ha automatizado completamente la búsqueda de matches con esos formatos. Se ha eliminado todas las clases TimeStampParser porque ya no son necesarias. Se ha cambiado el nombre del package.
