+++ GEN_TimeStamp/bin/.gitignore
@@ -1,1 +1,3 @@
 /main/
+/utils/
+/com/
