+++ GEN_TimeStamp/src/com/koldogontzal/timestamp/EjemploDeUso.java
@@ -1,32 +1,27 @@
-package main;
-
-import java.io.File;
+package com.koldogontzal.timestamp;
 
 public class EjemploDeUso {
 
 	public static void main(String[] args) {
 		// Ejemplo de uso de MarcaDeTiempo
-		String s = "2015.03.03";
+		String s = "xxxx.03.03 12_45_xx";
 		System.out.println(s);
 
 		TimeStamp m;
 		try {
 			// Crear una a partir de un String
 			m = new TimeStamp(s);
-			long l = new File("").lastModified();
-			System.out.println(l);
-			m = new TimeStamp(l);
+			//long l = new TimeStampFile("").lastModified();
+			//System.out.println(l);
+			//m = new TimeStamp(l);
 
 			// Aplicar distintos formatos
-			System.out.println(m);
-			System.out.println(m.toString(TimeStamp.FormatYYYY_MM_DD_hh_mm_ss));
-			System.out.println(m.toString(TimeStamp.FormatYYYYMMDD_hhmmss));
-			System.out.println(m.toString(TimeStamp.FormatYYYYMMDDhhmmss));
-			System.out.println(m.toString(TimeStamp.FormatYYYYMMDD_hhmm));
-			System.out.println(m.toString(TimeStamp.FormatYYYYMMDDhhmm));
-			System.out.println(m.toString(TimeStamp.FormatYYYY_MM_DD));
-			System.out.println(m.toString(TimeStamp.FormatYYYYMMDD));
-
+			System.out.println("Default: " + m);
+			TimeStampFormat [] listadoFormatos = TimeStampFormat.values();
+			for (TimeStampFormat f : listadoFormatos) {
+				System.out.println(f.name() + ": " + m.toString(f));
+			}
+			
 			// Modificar la fecha de una MarcaDeTiempo
 			long añadirUnDia = 1000 * (60 * 60 * 24 + 1);
 			m.addTime(añadirUnDia);
@@ -43,7 +38,7 @@ public class EjemploDeUso {
 		TimeStampFile f = new TimeStampFile(path);
 		System.out.println(f);
 
-		f.setTimeStamp(TimeStamp.FormatYYYY_MM_DD_hh_mm_ss);
+		f.setTimeStamp(TimeStampFormat.yyyy_MM_dd_HH_mm_ss);
 		System.out.println("\n\n" + f);
 
 	}
